+++ src/test/persistence/JsonTest.java
@@ -6,7 +6,7 @@ import model.Expense;
 public class JsonTest {
     //the following methods are partially credited to JsonSerializationDemo
 
-    protected void checkExpense(String name, Double cost, String category, Expense expense) {
+    protected void checkExpense(String name, Integer cost, String category, Expense expense) {
         assertEquals(name, expense.getName());
         assertEquals(category, expense.getCategory());
         assertEquals(cost, expense.getCost());
